+++ src/pages/TodoPage.tsx
@@ -1,13 +1,12 @@
-import { FormEvent, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useTodo } from '../contexts/TodoContext';
 import TodoItem from '../components/TodoItem';
 import { Todo } from '../types/Todo';
-import useInput from '../hooks/useInput';
+import { AddTodo } from '../components/AddTodo';
 
 const TodoPage = () => {
-  const { data: content, updateData: setContent } = useInput('');
   const [data, setData] = useState<Todo[]>([]);
-  const { getTodos, createTodo } = useTodo();
+  const { getTodos } = useTodo();
 
   useEffect(() => {
     getTodos().then((data) => {
@@ -15,19 +14,9 @@ const TodoPage = () => {
     });
   }, []);
 
-  const handleSubmit = (event: FormEvent) => {
-    event.preventDefault();
-    createTodo(content).then((newTodo) => {
-      setData([...data, newTodo]);
-    });
-  };
-
   return (
     <div>
-      <form onSubmit={handleSubmit}>
-        <input data-testid='new-todo-input' onChange={setContent} />
-        <button data-testid='new-todo-add-button'>만들기</button>
-      </form>
+      <AddTodo />
       {data.map((todoItem) => {
         return <TodoItem key={todoItem.id} data={todoItem} />;
       })}
