+++ src/reudx/store.ts
@@ -2,5 +2,4 @@ import { createStore } from 'redux';
 import { reducer } from './reducer';
 
 const store = createStore(reducer);
-
 export default store;
